==> 02 - Centers Capacity Tool - Preprocess Parcels.ipynb <==
import pandas as pd
import os
import numpy as np
import arcpy

# Import the required ArcGIS API for Python modules
import arcgis
from arcgis.gis import GIS

#get root folder
dir_working = os.getcwd()
print(dir_working)

#define folders
dir_process  = os.path.join(dir_working, "intermediate")
dir_results  = os.path.join(dir_working, "results"     )
dir_inputs   = os.path.join(dir_working, "input"       )
dir_data     = os.path.join(dir_inputs , "REMMdata"    )

BaseParcels        = os.path.join(dir_data  , r"parcels\Center01102019.gdb\BaseGP01102019"  )
Buildings          = os.path.join(dir_data  , r"BASE_run_285_2015_parcel_dev.csv"           )

ParcelsGDB = "parcels.gdb"

#name of new data features
Parcels     = os.path.join(dir_process, ParcelsGDB + "\Parcels"              )

#name of new data features
Parcels_shp = os.path.join(dir_process, "Parcels.shp"              )

def deleteIfExists(obj):
    if arcpy.Exists(obj): arcpy.Delete_management(obj)
        
#if processing geodatabase doesn't exist, create it
print("Checking if " + ParcelsGDB + " exists...")
if not arcpy.Exists(os.path.join(dir_process, ParcelsGDB)):
    print("Creating " + ParcelsGDB + "...")
    arcpy.management.CreateFileGDB(dir_process, ParcelsGDB)
else:
    print(ParcelsGDB + " exists...")


# Prep Parcels

#read in parcels
sdf_bPar = pd.DataFrame.spatial.from_featureclass(BaseParcels)

sdf_bPar

#set field for commercial, residential, or mixed

sdf_bPar['luType'] = ''

#undevelopable land (all building types are 'f')
sdf_bPar.loc[((sdf_bPar['type1'] == 'f') &
              (sdf_bPar['type2'] == 'f') &
              (sdf_bPar['type3'] == 'f') &
              (sdf_bPar['type4'] == 'f') &
              (sdf_bPar['type5'] == 'f') &
              (sdf_bPar['type6'] == 'f') &
              (sdf_bPar['type7'] == 'f') &
              (sdf_bPar['type8'] == 'f')), 'luType'] = 'Undevelopable'

sdf_bPar = sdf_bPar.drop(columns=['OBJECTID'      ,
                                  'city'          ,
                                  'x'             ,
                                  'y'             ,
                                  'county_name'   ,
                                  'distsml_id'    ,
                                  'distmed_id'    ,
                                  'distlrg_id'    ,
                                  'county_taz_id' ,
                                  'NoBuild'       ,
                                  'redev_friction',
                                  'Agriculture'   ,
                                  'Sliver'        ,
                                  'Split'         ,
                                  'utmxi'         ,
                                  'utmyi'         ,
                                  'note'          ,
                                  'zone_id'       ,
                                  'type1'         ,
                                  'type2'         ,
                                  'type3'         ,
                                  'type4'         ,
                                  'type5'         ,
                                  'type6'         ,
                                  'type7'         ,
                                  'type8'])

sdf_bPar

sdf_bPar_luTypeEmpty = sdf_bPar.loc[(sdf_bPar['luType'] == '')]
sdf_bPar_luTypeEmpty

sdf_bPar_lutypeTotals = sdf_bPar.groupby(['luType']).agg({"parcel_id": [np.size],"parcel_acre": [np.sum]})
sdf_bPar_lutypeTotals

sdf_bParProcessed = sdf_bPar[['parcel_id_REMM','parcel_acre','luType','SHAPE']]

sdf_bParProcessed

df_parcelwithbldg = pd.read_csv(Buildings)

display(Buildings)

pd.set_option('display.max_columns', None)
df_parcelwithbldg

df_bldg = df_parcelwithbldg[df_parcelwithbldg['has_buildings']==1].copy()

df_bldg['bldgtype'] = ''

#residential building types
df_bldg.loc[(df_bldg['is_sf'] != 0) | (df_bldg['is_mf'] != 0), 'bldgtype'] = 'Residential'

#commercial building types, if mixed use (res and com) then use commercial 
df_bldg.loc[((df_bldg['is_industrial'] != 0) |
             (df_bldg['is_retail'    ] != 0) |
             (df_bldg['is_office'    ] != 0) |
             (df_bldg['is_govt'      ] != 0) |
             (df_bldg['is_mixeduse'  ] != 0) |
             (df_bldg['is_other'     ] != 0)), 'bldgtype'] = 'Commercial'

df_bldg

df_bldg[df_bldg['parcel_id']==77033]

#"bldgtype":[np.min] --> 'Commercial' < 'Residential', pick commercial if a mix of types

df_bldg_summary = df_bldg.groupby(['parcel_id'], as_index=False).agg({"bldgtype":[np.min], "year_built":[np.max], "residential_units":[np.sum], "job_spaces":[np.sum], "residential_sqft":[np.sum], "non_residential_sqft":[np.sum], "res_price_per_sqft":[np.mean], "unit_price_non_residential":[np.mean]})
df_bldg_summary.columns = df_bldg_summary.columns.droplevel(1)

df_bldg_summary['value_res'] = df_bldg_summary['residential_sqft'] * df_bldg_summary['res_price_per_sqft']
df_bldg_summary['value_com'] = df_bldg_summary['non_residential_sqft'] * df_bldg_summary['unit_price_non_residential']

df_bldg_summary.columns = ('parcel_id','bldgtype','yearbuilt','resunits','job_spaces','sf_res','sf_com','sfvalue_res','sfvalue_com','value_res','value_com')

df_bldg_summary.to_csv(os.path.join(dir_process, "bldgsummary.csv"))
df_bldg_summary

sdf_bPar_bldg = pd.DataFrame.merge(sdf_bPar,df_bldg_summary,on='parcel_id',how='left')
sdf_bPar_bldg

#sdf_bPar_cParMax_bldg.sum()

print ("Exporting intermediate Parcel feature class...")
sdf_bPar_bldg.spatial.to_featureclass(location=Parcels)

print ("Calculating parcel acreage...")

# Execute DeleteField
arcpy.DeleteField_management(Parcels, ['Acres'])

arcpy.AddField_management(Parcels, "ParcelAcres" , "DOUBLE")
arcpy.CalculateField_management(Parcels,
                                "ParcelAcres",
                                "!shape.area@acres!",
                                "PYTHON_9.3")
print ("Done")

#clear memory
del sdf_bPar_bldg
del sdf_bPar
del df_bldg
del df_bldg_summary
