==> 03 - Centers Capacity Tool - Setup Analysis Areas.ipynb <==
import pandas as pd
import os
import numpy as np
import matplotlib.ticker as mtick
import arcpy

# Import the required ArcGIS API for Python modules
import arcgis
from arcgis.gis import GIS
from arcgis.geoanalytics import manage_data

from IPython.display import display, Markdown

#get root folder
dir_working = os.getcwd()
print(dir_working)

#define folders
dir_process  = os.path.join(dir_working, "intermediate")
dir_results  = os.path.join(dir_working, "results"     )
dir_inputs   = os.path.join(dir_working, "input"       )
dir_data     = os.path.join(dir_inputs , "REMMdata"    )

AnalysisAreasInput = os.path.join(dir_inputs, r"Parcel-Redevelopment.gdb\CentersTOD_ResidentialCapacity_Areas")
ClassParameters    = os.path.join(dir_inputs, r"class_parameters.csv"                       )

ParcelsGDB = "parcels.gdb"
Parcels    = os.path.join(dir_process, ParcelsGDB + "\Parcels"              )

ProcessGDB = "process.gdb"

#name of new data features
AnalysisAreas         = os.path.join(dir_process, ProcessGDB + "\AnalysisAreas"        )
AnalysisAreas_ByName  = os.path.join(dir_process, ProcessGDB + "\AnalysisAreas_ByName" )
AnalysisAreas_ByClass = os.path.join(dir_process, ProcessGDB + "\AnalysisAreas_ByClass")
AnalysisAreas_IDs     = os.path.join(dir_process, ProcessGDB + "\AnalysisAreas_IDs"    )
ParcelsAA_join        = os.path.join(dir_process, ProcessGDB + "\ParcelsAA_join"       )
ParcelsAA             = os.path.join(dir_process, ProcessGDB + "\ParcelsAA"            )

#name of new data features
AnalysisAreas_shp         = os.path.join(dir_process, "AnalysisAreas.shp"        )
AnalysisAreas_ByName_shp  = os.path.join(dir_process, "AnalysisAreas_ByName.shp" )
AnalysisAreas_ByClass_shp = os.path.join(dir_process, "AnalysisAreas_ByClass.shp")
AnalysisAreas_IDs_shp     = os.path.join(dir_process, "AnalysisAreas_IDs.shp"    )
Parcels_shp               = os.path.join(dir_process, "Parcels.shp"              )
ParcelsAA_shp             = os.path.join(dir_process, "ParcelsAA.shp"            )

def deleteIfExists(obj):
    if arcpy.Exists(obj): arcpy.Delete_management(obj)

print(AnalysisAreasInput)

# Prep Analysis Areas

#Create Analysis Area Layer in Intermediate Files and Fill with a AAName that combines Name and TOD status

sAreaBlankName  = 'No Area Name'

sAAName  = "AnalysisAreaName" #field for calculated Analysis Area Name
sAAClass = "AnalysisAreaClass" #field for calculated Analysis Area Class

#if processing geodatabase doesn't exist, create it
print("Checking if " + ProcessGDB + " exists...")
if not arcpy.Exists(os.path.join(dir_process, ProcessGDB)):
    print("Creating " + ProcessGDB + "...")
    arcpy.management.CreateFileGDB(dir_process, ProcessGDB)
else:
    print(ProcessGDB + " exists...")

print ("Creating Analysis Input Area intermediate layer in " + ProcessGDB + "...")
deleteIfExists(AnalysisAreas)
arcpy.management.Copy(AnalysisAreasInput, AnalysisAreas)

## Subroutines

print ("Adding " + sAAName + " field to " + AnalysisAreas + "...")
arcpy.AddField_management(AnalysisAreas, sAAClass, "TEXT", 50)

areaname_codeblock = """def createAreaName(centername):
    return centername"""

#create calculated Class for analysis area combining center Class and tod status
centerClass_codeblock = """def createClassName(centerClass):
   return centerClass"""


print ("Calculating " + sAAName + " field...")
arcpy.CalculateField_management(AnalysisAreas,
                                sAAName,
                                "createAreaName(!ClassCode!)",
                                "PYTHON_9.3",
                                areaname_codeblock)
print ("Calculating " + sAAClass + " field...")
arcpy.CalculateField_management(AnalysisAreas,
                                sAAClass,
                                "createClassName(!ClassCode!)",
                                "PYTHON_9.3",
                                centerClass_codeblock)

print ("Creating " + AnalysisAreas_ByName + " intermediate layer in " + ProcessGDB + "...")
deleteIfExists(AnalysisAreas_ByName)
arcpy.management.Dissolve(AnalysisAreas, AnalysisAreas_ByName,  sAAName)

print ("Creating " + AnalysisAreas_ByClass + " intermediate layer in " + ProcessGDB + "...")
deleteIfExists(AnalysisAreas_ByClass)
arcpy.management.Dissolve(AnalysisAreas, AnalysisAreas_ByClass, sAAClass)

print ("All Done")


#Areas data
df_AnalysisAreas = pd.DataFrame()

cursorName = arcpy.SearchCursor(AnalysisAreas_ByName)
row = cursorName.next()
while row:
    #print(row.getValue(sAAName))
    df_AnalysisAreas = df_AnalysisAreas.append({sAAName : row.getValue(sAAName)},ignore_index=True)
    row = cursorName.next()  
    
df_AnalysisAreas.index.name = 'AreaID'
df_AnalysisAreas = df_AnalysisAreas.reset_index()
df_AnalysisAreas.to_csv(os.path.join(dir_results, r'areas.csv'),index=False)
display(df_AnalysisAreas)

df_ClassParam = pd.read_csv(ClassParameters)
display(df_ClassParam)

sdf_AA

#create layer with just IDs to make union cleaner

df_ClassIDs = df_ClassParam[['ClassID','ClassDescription']]
#display(df_ClassIDs)

sdf_AA = pd.DataFrame.spatial.from_featureclass(AnalysisAreas)
#display(sdf_AnalysisAreas)

sdf_AAwClassID =  pd.DataFrame.merge(sdf_AA,df_ClassIDs,left_on='ClassCode', right_on='ClassID')
#display(sdf_AAwID)

sdf_AAwClassIDwNameID =  pd.DataFrame.merge(sdf_AAwClassID,df_AnalysisAreas,on=sAAName)
#display(sdf_AAwClassIDwNameID)

sdf_AAIDs = sdf_AAwClassIDwNameID[['AreaID','ClassID','SHAPE']]
#display(sdf_AAIDs)

#arcpy.Delete_management(AnalysisAreas_IDs)
deleteIfExists(AnalysisAreas_IDs)
sdf_AAIDs.spatial.to_featureclass(location=AnalysisAreas_IDs)
print('All Done')

sdf_AAwClassIDwNameID

# Intersect Analysis Areas with Parcels

deleteIfExists(ParcelsAA_join)

#intersection parcels with analysis area types
arcpy.analysis.Intersect([AnalysisAreas_IDs,Parcels], ParcelsAA_join)



print ("Calculating area...")
arcpy.AddField_management(ParcelsAA_join, "piece_acres" , "DOUBLE")
arcpy.CalculateField_management(ParcelsAA_join,
                                "PieceAcres",
                                "!shape.area@acres!",
                                "PYTHON_9.3")
arcpy.AddField_management(ParcelsAA_join, "piece_portion" , "DOUBLE")
arcpy.CalculateField_management(ParcelsAA_join,
                                "PiecePortion",
                                "$feature.PieceAcres / $feature.ParcelAcres", "ARCADE")

print ("Done")

sdf_ParcelsAA = pd.DataFrame.spatial.from_featureclass(ParcelsAA_join)
sdf_ParcelsAA.columns

#adjust parcel values by portion of parcel in parcel piece
sdf_ParcelsAA['parcel_sqft'] = sdf_ParcelsAA['parcel_sqft'] * sdf_ParcelsAA['piece_portion']
sdf_ParcelsAA['resunits'   ] = sdf_ParcelsAA['resunits'   ] * sdf_ParcelsAA['piece_portion']
sdf_ParcelsAA['job_spaces' ] = sdf_ParcelsAA['job_spaces' ] * sdf_ParcelsAA['piece_portion']
sdf_ParcelsAA['sf_res'     ] = sdf_ParcelsAA['sf_res'     ] * sdf_ParcelsAA['piece_portion']
sdf_ParcelsAA['sf_com'     ] = sdf_ParcelsAA['sf_com'     ] * sdf_ParcelsAA['piece_portion']
sdf_ParcelsAA['value_res'  ] = sdf_ParcelsAA['value_res'  ] * sdf_ParcelsAA['piece_portion']
sdf_ParcelsAA['value_com'  ] = sdf_ParcelsAA['value_com'  ] * sdf_ParcelsAA['piece_portion']

#make sure acres used in future is just for piece of parcel, not all
sdf_ParcelsAA['acres'] = sdf_ParcelsAA['piece_acres']
display(sdf_ParcelsAA)

display("exporting Parcels")

sdf_ParcelsAA.spatial.to_featureclass(location=ParcelsAA)

sdf_ParcelsAA.groupby(['area_id','class_id'], as_index=False).agg({"parcel_id": [np.size], "job_spaces": [np.sum], "resunits":[np.sum], "acres":[np.sum]})

#export to shp files for better drawing in Jupyter

deleteIfExists(AnalysisAreas_shp)
print ("Exporting " + AnalysisAreas_shp + "...")
arcpy.conversion.FeatureClassToShapefile(AnalysisAreas, dir_process)

deleteIfExists(AnalysisAreas_IDs_shp)
print ("Exporting " + AnalysisAreas_IDs_shp + "...")
arcpy.conversion.FeatureClassToShapefile(AnalysisAreas_IDs, dir_process)

deleteIfExists(ParcelsAA_shp)
print ("Exporting " + ParcelsAA_shp + "...")
arcpy.conversion.FeatureClassToShapefile(ParcelsAA, dir_process) 

display("done!")
