==> 04 - Centers Capacity Tool - Eligibility.ipynb <==
import pandas as pd
import os
import numpy as np
import matplotlib.ticker as mtick
import arcpy

# Import the required ArcGIS API for Python modules
import arcgis
gis = arcgis.GIS()

from IPython.display import display, Markdown

#scenario name
#allow change of scen_name
#scenario names must have an associated class parameters file in the config folder
scen_names = ["_55","_70","_85","_orig"]

#get root folder
dir_working = os.getcwd()
print(dir_working)

#define folders
dir_data     = os.path.join(dir_working, "data"        )
dir_config   = os.path.join(dir_working, "config"      )
dir_process  = os.path.join(dir_working, "intermediate")
dir_results  = os.path.join(dir_working, "results"     )
dir_inputs   = os.path.join(dir_working, "input"       )
dir_webapp   = os.path.join(dir_working, "forWebApp"   )

ProcessGDBname = "process.gdb"
FinalGDBname   = "parcel_redevelopment_potential.gdb"

ProcessGDBpath = os.path.join(dir_process, ProcessGDBname)
FinalGDBpath   = os.path.join(dir_results, FinalGDBname  )

ParcelsDevCategory = os.path.join(dir_process, ProcessGDB + "\ParcelsAA_DevCategory" + scen_name)
ParcelsWebApp      = os.path.join(dir_webapp , WebAppGDB  + "\AAParcels"             + scen_name)
DevCatWebApp       = os.path.join(dir_webapp , WebAppGDB  + "\AADevCategories"       + scen_name)

#name of new data features
AnalysisAreas      = os.path.join(dir_process, "analysisareas.shp"        )
AnalysisAreas_IDs  = os.path.join(dir_process, "analysisareas_ids.shp"    )
ParcelsAA          = os.path.join(dir_process, "parcelsaa.shp"            )
AnalysisAreasTable = os.path.join(dir_results, "areas.csv"                )

def deleteIfExists(obj):
    if arcpy.Exists(obj): arcpy.Delete_management(obj)
        
                
#if processing geodatabase doesn't exist, create it
print("Checking if " + FinalGDBname + " exists...")
if not arcpy.Exists(FinalGDBpath):
    print("Creating " + FinalGDBname + "...")
    arcpy.management.CreateFileGDB(dir_results, FinalGDBname)
else:
    print(FinalGDBname + " exists...")

print("done")

AnalysisAreas

sdf_AA = pd.DataFrame.spatial.from_featureclass(AnalysisAreas)
sdf_AA

# Parcels

sdf_ParcelsAA = pd.DataFrame.spatial.from_featureclass(ParcelsAA)
sdf_ParcelsAA.groupby(['area_id','class_id'], as_index=False).agg({"parcel_id": [np.size], "job_spaces": [np.sum], "resunits":[np.sum], "acres":[np.sum]})

sdf_ParcelsAA['county_id']

# Land Use

display(sdf_ParcelsAA.columns)
sdf_ParcelsAA = sdf_ParcelsAA.rename(columns={"lu_type": "lu_type_ex"})
display(sdf_ParcelsAA.columns)

sdf_ParcelsAA.loc[(sdf_ParcelsAA['lu_type_ex'] != "Undevelopable") & (sdf_ParcelsAA['bldgtype'] == 'Residential'), 'lu_type_ex'] = "Residential"
sdf_ParcelsAA.loc[(sdf_ParcelsAA['lu_type_ex'] != "Undevelopable") & (sdf_ParcelsAA['bldgtype'] == 'Commercial' ), 'lu_type_ex'] = "Commercial"
sdf_ParcelsAA.loc[(sdf_ParcelsAA['lu_type_ex'] != "Undevelopable") & ((sdf_ParcelsAA['bldgtype'] != 'Residential') & (sdf_ParcelsAA['bldgtype'] != 'Commercial')), 'lu_type_ex'] = "Developable"

# Building Age

sdf_ParcelsAA.columns

#set building age with base year 2019
sdf_ParcelsAA['BldAge'] = 0
sdf_ParcelsAA.loc[((sdf_ParcelsAA['yearbuilt'] > 0)), 'bld_age'] = 2019 - sdf_ParcelsAA['yearbuilt']
display(sdf_ParcelsAA)
display(sdf_ParcelsAA.columns)

# Parcel Value

sdf_ParcelsAA.columns

#set building age with base year 2019
sdf_ParcelsAA['ValuePerAcre'] = 0
sdf_ParcelsAA.loc[((sdf_ParcelsAA['lu_type_ex'] == 'Residential')), 'ValuePerAcre'] = sdf_ParcelsAA['value_res'] / sdf_ParcelsAA['acres']
sdf_ParcelsAA.loc[((sdf_ParcelsAA['lu_type_ex'] == 'Commercial' )), 'ValuePerAcre'] = sdf_ParcelsAA['value_com'] / sdf_ParcelsAA['acres']

#classify existing acreage by existing land use

sdf_ParcelsAA['acres_res'] = 0.0
sdf_ParcelsAA['acres_com'] = 0.0

sdf_ParcelsAA.loc[((sdf_ParcelsAA['lu_type_ex'] == 'Residential')), 'acres_res'] = sdf_ParcelsAA['acres']
sdf_ParcelsAA.loc[((sdf_ParcelsAA['lu_type_ex'] == 'Commercial' )), 'acres_com'] = sdf_ParcelsAA['acres']

display(sdf_ParcelsAA)

# Development Categories

scen_names

for scen_name in scen_names:
    
    display ('Processing ' + scen_name + ". Calculating parcel eligibility.")
    ClassParametersScenario = os.path.join(dir_config, r"class_parameters" + scen_name + ".csv")

    df_ClassParam = pd.read_csv(ClassParametersScenario)
    #display(df_ClassParam)
    
    
    #ReDevelop, NewDevelop, NoDevelop

    sdf_ParcelsAAwParam = pd.DataFrame.merge(sdf_ParcelsAA,df_ClassParam,left_on='class_id',right_on="ClassID",how='left')
    #display(sdf_ParcelsAAwParam)

    #undeveloped
    sdf_ParcelsDevCat = sdf_ParcelsAAwParam.copy()
    sdf_ParcelsDevCat['DevCategory'] = 'Remain'
    sdf_ParcelsDevCat.loc[(sdf_ParcelsDevCat['lu_type_ex'] == 'Developable'), 'DevCategory'] = "Developable"

    #initialize column
    sdf_ParcelsDevCat['DevCategoryAge'] = ''

    #check if redevelopable - step 1: building age
    #residential
    sdf_ParcelsDevCat.loc[(sdf_ParcelsDevCat['bld_age'] >= sdf_ParcelsDevCat['RedevBldgAgeRes_Low']) & (sdf_ParcelsDevCat['BldAge'] <= sdf_ParcelsDevCat['RedevBldgAgeRes_High']) & (sdf_ParcelsDevCat['lu_type_ex'] == 'Residential'), 'DevCategoryAge'] = "Redevelopable"
    #non-residential
    sdf_ParcelsDevCat.loc[(sdf_ParcelsDevCat['bld_age'] >= sdf_ParcelsDevCat['RedevBldgAgeCom_Low']) & (sdf_ParcelsDevCat['BldAge'] <= sdf_ParcelsDevCat['RedevBldgAgeCom_High']) & (sdf_ParcelsDevCat['lu_type_ex'] != 'Residential'), 'DevCategoryAge'] = "Redevelopable"

    #initialize column
    sdf_ParcelsDevCat['DevCategoryValue'] = ''

    #check if redevelopable - step 2: value
    #residential
    sdf_ParcelsDevCat.loc[(sdf_ParcelsDevCat['ValuePerAcre'] > 0) & (sdf_ParcelsDevCat['ValuePerAcre'] <= sdf_ParcelsDevCat['RedevValuePerAcreRes']) & (sdf_ParcelsDevCat['lu_type_ex'] == 'Residential'), 'DevCategoryValue'] = "Redevelopable"
    #commercial
    sdf_ParcelsDevCat.loc[(sdf_ParcelsDevCat['ValuePerAcre'] > 0) & (sdf_ParcelsDevCat['ValuePerAcre'] <= sdf_ParcelsDevCat['RedevValuePerAcreCom']) & (sdf_ParcelsDevCat['lu_type_ex'] == 'Commercial' ), 'DevCategoryValue'] = "Redevelopable"

    #depending on AND or OR defenition, determine whether a parcel is eligible for redevelopment
    sdf_ParcelsDevCat.loc[ (sdf_ParcelsDevCat['RedevAndOr'] =='AND') &  (sdf_ParcelsDevCat['DevCategoryAge'] == 'Redevelopable') & (sdf_ParcelsDevCat['DevCategoryValue'] == 'Redevelopable')  , 'DevCategory'] = "Redevelopable"
    sdf_ParcelsDevCat.loc[((sdf_ParcelsDevCat['RedevAndOr'] =='OR' ) & ((sdf_ParcelsDevCat['DevCategoryAge'] == 'Redevelopable') | (sdf_ParcelsDevCat['DevCategoryValue'] == 'Redevelopable'))), 'DevCategory'] = "Redevelopable"

    #undevelopable parcels
    sdf_ParcelsDevCat.loc[(sdf_ParcelsDevCat['lu_type_ex'] == 'Undevelopable'), 'DevCategory'] = "Undevelopable"


    #single-family redevelopment additional criteria (redeveloped parcel must include enough area for at least 1.0 more unit than original)

    #calculate redev potential for SF units... resunits range from 0.95 to 1.05 to account for rounding error
    sdf_ParcelsDevCat.loc[(sdf_ParcelsDevCat['DevCategory'] == 'Redevelopable') & (sdf_ParcelsDevCat['SFRedevFullAdd'] =='Yes') & (sdf_ParcelsDevCat['resunits'] > 0.95) & (sdf_ParcelsDevCat['resunits'] < 1.05), 'SFRedevAdd'] = sdf_ParcelsDevCat['CapacityRes_DUA'] * (sdf_ParcelsDevCat['acres'] * (1-sdf_ParcelsDevCat['PercentOpenSpace'])) - sdf_ParcelsDevCat['resunits']

    #change to 'Remain' for SF parcels with not enough redev space for additional units
    sdf_ParcelsDevCat.loc[(sdf_ParcelsDevCat['DevCategory'] == 'Redevelopable') & (sdf_ParcelsDevCat['SFRedevFullAdd'] =='Yes') & (sdf_ParcelsDevCat['resunits'] > 0.95) & (sdf_ParcelsDevCat['resunits'] < 1.05) & (sdf_ParcelsDevCat['SFRedevAdd'] < 1.0), 'DevCategory']      = 'Remain'

    #changed to Remain
    sdf_ParcelsDevCat.loc[(sdf_ParcelsDevCat['SFRedevAdd'] < 1.0), 'DevCategory_Flag'] = 'SFRedevAdd_NoPotential'

    #create layer names for intermediate feature classes
    ParcelsDevCategory          = os.path.join(ProcessGDBpath, "ParAA_DevCat" + scen_name)
    ParcelsDevCategory_simple   = os.path.join(ProcessGDBpath, "ParAA_DevCat" + scen_name + "_simple")
    
    #create layer names for final feature class
    ParcelsDevCategory_final    = os.path.join(FinalGDBpath,"Parcel_Redevelopment_Potential" + scen_name)
    
    deleteIfExists(ParcelsDevCategory)
    deleteIfExists(ParcelsDevCategory_simple)
    deleteIfExists(ParcelsDevCategory_final)

    #limit to only needed fields
    sdf_ParcelsDevCat = sdf_ParcelsDevCat[['parcel_id','DevCategory', 'DevCategoryAge',
           'DevCategoryValue', 'SFRedevAdd', 'DevCategory_Flag','SHAPE']]

    #rename columns -- sometimes crazy if column names are "sanitized"
    sdf_ParcelsDevCat.columns = ('parcel_id','dev_cat', 'dev_cat_age',
           'dev_cat_value', 'sf_redev_add', 'dev_cat_flag','SHAPE')
    
    #export to GeoDB
    display("exporting " + ParcelsDevCategory)

    #export first layer
    sdf_ParcelsDevCat.spatial.to_featureclass(location=ParcelsDevCategory)
    
    display("simplyfing " + scen_name)
    
    #simplify layer... removes some polygon vertices... parcels layer is so large, this reduces file size quite a bit
    #https://pro.arcgis.com/en/pro-app/latest/tool-reference/cartography/simplify-polygon.htm
    arcpy.cartography.SimplifyPolygon(ParcelsDevCategory, ParcelsDevCategory_simple, 'POINT_REMOVE', 1)

    display("dissolving " + scen_name)
    #dissovle, but creating dissolved layer, the .dbf portion of shapefile is super small
    arcpy.management.Dissolve(ParcelsDevCategory_simple, ParcelsDevCategory_final, ["dev_cat"])

    display(ParcelsDevCategory_final + " is done")
    
display('All Done!')

sdf_ParcelsDevCat

display(df_ClassParam)
df_ClassParam.dtypes

#aggregate parcels into analysis areas / classes
df_AADevCat = sdf_ParcelsDevCat.groupby(['area_id','class_id','county_id','dev_category'],as_index=False).agg({"parcel_id": [np.size], "job_spaces": [np.sum], "resunits":[np.sum], "acres":[np.sum], "sf_res":[np.sum], "sf_com":[np.sum], "acres_res":[np.sum], "acres_com":[np.sum]})
df_AADevCat.columns = df_AADevCat.columns.droplevel(1)
display(df_AADevCat)


df_AADevCat.to_csv(os.path.join(dir_results, r'AnalysisAreas_byDevCategory' + scen_name + '.csv'),index=False)
